[Stable] Refactor RBAC to use pundit gem and permissions

+++ app/controllers/api/v1x0/workflows_controller.rb
@@ -1,23 +1,20 @@
 module Api
   module V1x0
     class WorkflowsController < ApplicationController
-      include Api::V1x0::Mixins::IndexMixin
-      include Api::V1x0::Mixins::RBACMixin
-
-      before_action :create_access_check, :only => %i[create]
-      before_action :update_access_check, :only => %i[update]
-      before_action :destroy_access_check, :only => %i[destroy]
+      include Mixins::IndexMixin
 
       def create
+        authorize Workflow
+
         workflow = WorkflowCreateService.new(params.require(:template_id)).create(params_for_create)
         json_response(workflow, :created)
       end
 
-      # TODO: remove 'approval:workflows:read' from approver acls list in RBAC Insight
       def show
-        raise Exceptions::NotAuthorizedError, "Not Authorized for workflows" if Insights::API::Common::RBAC::Access.enabled? && !admin?
+        workflow = Workflow.find(params.require(:id))
+        authorize workflow
 
-        json_response(Workflow.find(params.require(:id)))
+        json_response(workflow)
       end
 
       def index
@@ -30,11 +27,13 @@ module Api
                      Workflow.all
                    end
 
-        collection(index_scope(relation))
+        collection(policy_scope(relation))
       end
 
       def destroy
         workflow = Workflow.find(params.require(:id))
+        authorize workflow
+
         workflow.destroy!
         head :no_content
       rescue ActiveRecord::InvalidForeignKey => e
@@ -46,12 +45,18 @@ module Api
       end
 
       def link
+        workflow = Workflow.find(params.require(:id))
+        authorize workflow
+
         WorkflowLinkService.new(params.require(:id)).link(params_for_create.to_unsafe_h)
 
         head :no_content
       end
 
       def unlink
+        workflow = Workflow.find(params.require(:id))
+        authorize workflow
+
         WorkflowUnlinkService.new(params[:id]).unlink(resource_object_params)
 
         head :no_content
@@ -59,6 +64,8 @@ module Api
 
       def update
         workflow = Workflow.find(params.require(:id))
+        authorize workflow
+
         # TODO: need to change params_for_update when using insights-api-commons
         WorkflowUpdateService.new(workflow.id).update(params_for_create)
 
@@ -66,13 +73,6 @@ module Api
       end
 
       private
-
-      # TODO: remove 'approval:workflows:read' from approver acls list in RBAC Insight
-      def rbac_scope(relation)
-        raise Exceptions::NotAuthorizedError, "Not Authorized for #{relation.model.table_name}" unless admin?
-
-        relation
-      end
 
       def collection(base_query)
         resp = Insights::API::Common::PaginatedResponse.new(
